requirements.txt
click
nrel-rex
numpy
pandas
pypistats
requests
seaborn
